Fix csv journaling failures
Starting 5 day ago, the nightly runs of the subwaydata.nyc ETL pipeline
have been failing to create csvs (more precisely: creating csvs with no trips).
This turns out to be a bug in the GTFS Go package that was introduced in a
commit 2 years ago! This is the commit:
https://github.com/jamespfennell/gtfs/commit/7e0a9db5605b4bdf6c48f6e0fade7a2db2d017e6

The bug is that the NyctIsAssigned field is no longer being populated, so
the journaler thinks all trips are unassigned, and skips all of them. This
bug from 2 years only kicked in now because I redeployed subwaydata.nyc for
the first time in 2+ years this week.

The fix uses the fact that NyctIsAssigned=true means that the trip has an
associated vehicle. So instead of consulting the boolean, we just check for
the presense of a vehicle. I will change the GTFS package to remove the
NyctIsAssigned field.